fix seeding

+++ T1809E_Project_Sem3/Models/Cart.cs
@@ -10,7 +10,7 @@ namespace T1809E_Project_Sem3.Models
         public Dictionary<int, Cart> Items { get; set; }
         public Product Product { get; set; }
         public int Quantity { get; set; }
-        public Decimal TotalPrice { get; set; }
+        public double TotalPrice { get; set; }
 
 
         public Cart(Product product, int quantity)
